FAPI: rename files get_intl_cert to get_web_cert.
Because now also AMD certificates are supported these
files are renamed.
